Fixed a bug on alert sound, continued the implementation of TalkLogMonitor following a TDD approach.
Alert sound: an event was added to the timer every every time the auto
dial was called with Alt+F12. That - along with the high priority set on
Auto Dial process - caused the sound card to stop working after a while.
Forcing the headset to be removed and re-plugged (on the log it was
possible to see several 'Sound played' logs.