Fixed a bug in number recognition
In some job the string 'Autodial:' is present. Being uncapitalised the
'l' is replaced with a '1'. That along with the ':' interpreted as a
space mix everything up. I applied changes so that the number il looked
up after a space and a number is validated against the rule of being a
10 digits number.